update-6th

+++ src/java/controller/customer/RemoveFromCart.java
@@ -1,37 +1,44 @@
 package controller.customer;
 
 import dao.DAO;
+import entity.Account;
 import entity.Category;
 import java.io.IOException;
-import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import java.util.List;
 
 public class RemoveFromCart extends HttpServlet {
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-        String id = request.getParameter("cartid");
-        DAO dao = new DAO();
-        List<Category> listC = dao.getAllCategory();
-        request.setAttribute("listC", listC);
-        dao.RemoveFromCart(id);
-        response.sendRedirect("showcart");
-    } 
+        HttpSession session = request.getSession();
+        Account a = (Account) session.getAttribute("acc");
+        if (a != null) {
+            String id = request.getParameter("cartid");
+            DAO dao = new DAO();
+            List<Category> listC = dao.getAllCategory();
+            request.setAttribute("listC", listC);
+            dao.RemoveFromCart(id);
+            response.sendRedirect("showcart");
+        } else {
+            response.sendRedirect("Login.jsp");
+        }
+    }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
-    } 
+    }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
     }
