+++ src/java/controller/seller/DeleteProductControlServlet.java
@@ -1,34 +1,41 @@
-
 package controller.seller;
 
 import dao.DAO;
+import entity.Account;
 import java.io.IOException;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 public class DeleteProductControlServlet extends HttpServlet {
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-        String id = request.getParameter("pid");
-        DAO dao = new DAO();
-        dao.DeleteProduct(id);
-        response.sendRedirect("manager");
-    } 
+        HttpSession session = request.getSession();
+        Account a = (Account) session.getAttribute("acc");
+        if (a != null && a.getIsSeller() == 1) {
+            String id = request.getParameter("pid");
+            DAO dao = new DAO();
+            dao.DeleteProduct(id);         
+            response.sendRedirect("manager");
+        } else {
+            response.sendRedirect("Login.jsp");
+        }
 
+    }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
-    } 
+    }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
     }
